more directory reorganizing editor components to editor subdirectory created shared directory for shared components, for now there is just image.jsx
