+++ client/src/components/Editor/Segment.jsx
@@ -1,6 +1,6 @@
 import React, { useContext } from "react";
 import { Box, InputBase, useTheme } from "@mui/material";
-import MinutesContext from "../contexts/MinutesContext";
+import MinutesContext from "../../contexts/MinutesContext";
 
 function Segment({ segmentIndex }) {
   const theme = useTheme();
