+++ client/src/components/Editor/Segment.test.jsx
@@ -2,10 +2,10 @@ import React from "react";
 import { expect, test, describe, beforeEach, vi, afterEach } from "vitest";
 import { fireEvent, render, screen } from "@testing-library/react";
 import { ThemeProvider } from "@mui/material/styles";
-import theme from "../theme";
+import theme from "../../theme";
 import Segment from "./Segment";
-import MinutesContext from "../contexts/MinutesContext";
-import { mockMinutesContextState } from "../util/test.helpers";
+import MinutesContext from "../../contexts/MinutesContext";
+import { mockMinutesContextState } from "../../util/test.helpers";
 
 describe("Segment", () => {
   const updateMinutesMock = vi.fn();
