+++ client/src/components/Editor/SegmentButtons.jsx
@@ -3,7 +3,7 @@ import { Box, IconButton, useTheme } from "@mui/material";
 import KeyboardArrowUpIcon from "@mui/icons-material/KeyboardArrowUp";
 import KeyboardArrowDownIcon from "@mui/icons-material/KeyboardArrowDown";
 import DeleteIcon from "@mui/icons-material/Delete";
-import MinutesContext from "../contexts/MinutesContext";
+import MinutesContext from "../../contexts/MinutesContext";
 
 function SegmentButtons({ segmentIndex }) {
   const theme = useTheme();
