+++ client/src/components/Editor/SegmentButtons.test.jsx
@@ -2,10 +2,10 @@ import React from "react";
 import { expect, test, describe, beforeEach, vi, afterEach } from "vitest";
 import { render, screen, waitFor } from "@testing-library/react";
 import { ThemeProvider } from "@mui/material/styles";
-import theme from "../theme";
+import theme from "../../theme";
 import SegmentButtons from "./SegmentButtons";
-import MinutesContext from "../contexts/MinutesContext";
-import { mockMinutesContextState } from "../util/test.helpers";
+import MinutesContext from "../../contexts/MinutesContext";
+import { mockMinutesContextState } from "../../util/test.helpers";
 
 describe("EditorButtons", () => {
   const updateMinutesMock = vi.fn();
