+++ client/src/components/Editor/SegmentContainer.jsx
